make llama compile
