+++ lit_llama/model.py
@@ -19,6 +19,7 @@ class LLaMAConfig:
     n_layer: int = 32
     n_head: int = 32
     n_embd: int = 4096
+    complex_rope: bool = True
 
     @classmethod
     def from_name(cls, name: str) -> Self:
@@ -104,6 +105,7 @@ class CausalSelfAttention(nn.Module):
         self.n_head = config.n_head
         self.n_embd = config.n_embd
         self.block_size = config.block_size
+        self.complex_rope = config.complex_rope
         self.rope_cache = None
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
@@ -124,6 +126,7 @@ class CausalSelfAttention(nn.Module):
                 n_elem=self.n_embd // self.n_head, 
                 dtype=x.dtype,
                 device=x.device,
+                complex=self.complex_rope,
             )
 
         q = apply_rope(q, self.rope_cache)
@@ -188,13 +191,20 @@ class RMSNorm(nn.Module):
         return self.scale * x_normed
 
 
-def build_rope_cache(seq_len: int, n_elem: int, dtype: torch.dtype, device: torch.device, base: int = 10000) -> torch.Tensor:
+
+def build_rope_cache(seq_len: int, n_elem: int, dtype: torch.dtype, device: torch.device, base: int = 10000, complex: bool = False) -> torch.Tensor:
     """Enhanced Transformer with Rotary Position Embedding.
 
     Derived from: https://github.com/labmlai/annotated_deep_learning_paper_implementations/blob/master/labml_nn/
     transformers/rope/__init__.py. MIT License:
     https://github.com/labmlai/annotated_deep_learning_paper_implementations/blob/master/license.
     """
+    if complex:
+        from lit_llama.utils import build_rope_cache_complex
+
+        return build_rope_cache_complex(seq_len=seq_len, n_elem=n_elem, dtype=dtype, device=device, base=base)
+    
+
     # $\Theta = {\theta_i = 10000^{\frac{2(i-1)}{d}}, i \in [1, 2, ..., \frac{d}{2}]}$
     theta = 1.0 / (base ** (torch.arange(0, n_elem, 2, dtype=dtype, device=device) / n_elem))
 
@@ -204,31 +214,34 @@ def build_rope_cache(seq_len: int, n_elem: int, dtype: torch.dtype, device: torc
     # Calculate the product of position index and $\theta_i$
     idx_theta = torch.outer(seq_idx, theta)
 
-    # Compute cache. Because polar only takes float32 or float64, we need to cast
-    # when working with 16 bit floats (float16 or bfloat16)
-    working_dtype = (
-        torch.float32 if (dtype == torch.float16 or dtype == torch.bfloat16) else dtype
-    )
-    complex_dtype = (
-        torch.complex32
-        if (dtype == torch.float16 or dtype == torch.bfloat16)
-        else torch.complex64
-    )
-    cache = torch.polar(
-        torch.ones_like(idx_theta).to(working_dtype), idx_theta.to(working_dtype)
-    ).to(complex_dtype)
-    return cache
+    # Concatenate so that for row $m$ we have
+    # $[m \theta_0, m \theta_1, ..., m \theta_{\frac{d}{2}}, m \theta_0, m \theta_1, ..., m \theta_{\frac{d}{2}}]$
+    idx_theta2 = torch.cat([idx_theta, idx_theta], dim=1)
+
+    # Cache them
+    cos_cache = idx_theta2.cos()[None, None, :, :]
+    sin_cache = idx_theta2.sin()[None, None, :, :]
+
+    return torch.stack((cos_cache, sin_cache), dim=0)
 
 
 def apply_rope(x: torch.Tensor, rope_cache: torch.Tensor) -> torch.Tensor:
-    x = x.transpose(1, 2)
+    if rope_cache.dtype in (torch.complex32, torch.complex64):
+        from lit_llama.utils import apply_rope_complex
+        
+        return apply_rope_complex(x, rope_cache)
 
+    neg_half_x = rotate_neg_half(x)
+    cos, sin = rope_cache
     # truncate to support variable sizes
-    T = x.size(1)
-    rope_cache = rope_cache[:T]
+    T = x.size(2)
+    cos = cos[:, :, :T]
+    sin = sin[:, :, :T]
+    return (x * cos) + (neg_half_x * sin)
 
-    # cast because `view_as_complex` does not support 16 bit tensors
-    xc = torch.view_as_complex(x.float().reshape(*x.shape[:-1], -1, 2))
-    rope_cache = rope_cache.view(1, xc.size(1), 1, xc.size(3))
-    x_out = torch.view_as_real(xc * rope_cache).flatten(3)
-    return x_out.transpose(1, 2).type_as(x)
+
+def rotate_neg_half(x: torch.Tensor) -> torch.Tensor:
+    # $\frac{d}{2}$
+    d_2 = x.shape[-1] // 2
+    # Calculate $[-x^{(\frac{d}{2} + 1)}, -x^{(\frac{d}{2} + 2)}, ..., -x^{(d)}, x^{(1)}, x^{(2)}, ..., x^{(\frac{d}{2})}]$  # noqa: E501
+    return torch.cat([-x[:, :, :, d_2:], x[:, :, :, :d_2]], dim=-1)
